requirements.txt
Pillow
imagehash
tqdm
